SonarQube issue solved
